cctv_portal/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
ultralytics>=8.2.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe>=0.10.0
facenet-pytorch>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.5.0
python-multipart>=0.0.9
websockets>=12.0
boxmot>=10.0.0
